+++ packages/mg-components/.storybook/main.ts
@@ -124,6 +124,7 @@ const config: StorybookConfig = {
         'v5.28.1': 'https://626149b307606d003ada26b4-vgqkjpyvnp.chromatic.com',
         'v5.28.2': 'https://626149b307606d003ada26b4-abvittqvfe.chromatic.com',
         'v6.0.0': 'https://626149b307606d003ada26b4-atannsibmz.chromatic.com',
+        'v6.1.0': 'https://626149b307606d003ada26b4-oclwtefwcd.chromatic.com',
       },
     },
   },