+++ packages/mg-components/src/components/molecules/inputs/mg-input-checkbox/mg-input-checkbox.tsx
@@ -324,6 +324,19 @@ export class MgInputCheckbox implements Omit<MgInputCheckboxListProps, 'id' | 'c
   }
 
   /**
+   * Reset value, validity and error state
+   */
+  @Method()
+  async reset(): Promise<void> {
+    if (!this.readonly) {
+      this.value = this.value.map(item => ({ ...item, value: false }));
+      this.checkValidity();
+      this.errorMessage = undefined;
+      this.hasDisplayedError = false;
+    }
+  }
+
+  /**
    * Method to set validity values
    * @param newValue - valid new value
    */