+++ packages/mg-components/src/components/molecules/inputs/mg-input-text/mg-input-text.tsx
@@ -309,6 +309,19 @@ export class MgInputText {
   }
 
   /**
+   * Reset value, validity and error state
+   */
+  @Method()
+  async reset(): Promise<void> {
+    if (!this.readonly) {
+      this.value = '';
+      this.checkValidity();
+      this.errorMessage = undefined;
+      this.hasDisplayedError = false;
+    }
+  }
+
+  /**
    * Method to set validity values
    * @param newValue - valid new value
    */