+++ packages/mg-components/src/components/molecules/mg-form/mg-form.conf.ts
@@ -1,3 +1,5 @@
+import { AriaRole } from '../../../types';
+
 /**
  * List of all possibles roles
  */
@@ -6,7 +8,7 @@ export const roles = ['form', 'search', 'none', 'presentation'] as const;
 /**
  * Aria Role type from roles
  */
-export type AriaRoleType = (typeof roles)[number];
+export type AriaRoleType = AriaRole & (typeof roles)[number];
 
 /**
  * List of all possibles required message status