+++ packages/stencil-helpers/src/components/index.ts
@@ -171,6 +171,13 @@ export const toString = (value: unknown): string => {
 };
 
 /**
+ * Validate number
+ * @param value - value to check
+ * @returns `true` if number is valid
+ */
+export const isValidNumber = (value: unknown): value is number => typeof value === 'number' && !Number.isNaN(value);
+
+/**
  * Cleans string characters by removing special characters and converting to lowercase.
  * @param text - text to clean
  * @returns cleaned string